style:Home styled
Home page has been styled. There is a picture of the logo and name of website. Also added a image folder in the static folder. Did some debugging to get this branch up to date with the login branch since it had some styling work that needed to be in this branch.
GitHub Issue ##27